+++ tests/hazmat/primitives/test_hashes.py
@@ -4,6 +4,7 @@
 
 
 import binascii
+import sys
 import typing
 
 import pytest
@@ -252,6 +253,12 @@ class TestSHAKE:
 
         with pytest.raises(ValueError):
             xof(digest_size=0)
+
+    @pytest.mark.parametrize("xof", [hashes.SHAKE128, hashes.SHAKE256])
+    def test_xof(self, xof):
+        algorithm = xof.xof()
+        assert isinstance(algorithm, xof)
+        assert algorithm.digest_size == sys.maxsize
 
 
 @pytest.mark.supported(